Widcart support is back, overflows are hopefully fixed.

+++ src/dup.cpp
@@ -3,7 +3,7 @@
 map<string, int > DupDict;
 
 //Version with strings & substrings. Not computationally efficient.
-void screen_duplicates(Read *seq1, Read *seq2, unsigned long &duplicates) {
+void screen_duplicates(Read *seq1, Read *seq2, unsigned long long &duplicates) {
     int c = 0;
     
     if( (seq1->initial_length < start_dw+size_dw) && (seq2->initial_length < start_dw+size_dw) ) return;
